Found the Problem
Alex. The problem for the if statements was the fact that every
statement in a statment list had to be followed by a semicolon. I added
a new rule that allowed you to substitute a statement without a
semicolon at the very end of it. But I am running into a node error
when I try to traverse the tree. Please look at it when you have a
chance